some fixs and norminette, input merge

+++ includes/ft_string.h
@@ -17,12 +17,12 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-typedef struct s_string 			t_string;
+typedef struct s_string				t_string;
 struct s_string {
 	size_t	(*len)(const char *string);
 	char	**(*split)(char const *s, char c);
 	char	*(*n_str)(const char *s1, const char *s2, size_t n);
-	size_t	(*copy_n)(char *dest, const char *src, size_t size);
+	int		(*copy_n)(char *dest, const char *src, size_t size);
 	char	*(*join)(char const *s1, char const *s2);
 	int		(*only_this_chars)(char *s1, char *s2);
 	char	*(*pos_char)(const char *string, int c);
@@ -39,7 +39,7 @@ struct s_string {
 size_t		ft_strlen(const char *string);
 char		**ft_split(char const *s, char c);
 char		*ft_strnstr(const char *s1, const char *s2, size_t n);
-size_t		ft_strlcpy(char *dest, const char *src, size_t size);
+int			ft_strlcpy(char *dest, const char *src, size_t size);
 char		*ft_strdup(const char *s);
 int			ft_strncmp(const char *str1, const char *str2, size_t n);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
@@ -47,6 +47,6 @@ char		*ft_string_in_two_char(char *str, char c1, char c2);
 char		*replace_str(char *str1, char *str2);
 int			index_char(char *str, char c);
 char		*split_two_str(char *str, char c, int option);
-t_string 	*string(void);
+t_string	*string(void);
 
 #endif
